refactor(shared_kernel): promote LibraryId from library module
LibraryId is consumed by identity (profile library ACL) and media
(catalog filtering port) besides its owning module. Move it to the
shared_kernel with a back-compat re-export in library (ADR-018).
