+++ src/modules/library/domain/value_objects/library_id.py
@@ -1,25 +1,11 @@
-"""Library external ID value object."""
+"""Library external ID value object.
 
-from typing import ClassVar
+``LibraryId`` moved to the shared_kernel (ADR-018) because it is
+consumed by ``identity`` (profile library ACL) and ``media`` (catalog
+filtering port) in addition to this module. This re-export keeps the
+historical import path working while callers migrate incrementally.
+"""
 
-from src.building_blocks.domain.external_id import ExternalId
-
-
-class LibraryId(ExternalId):
-    """External ID for libraries.
-
-    Format: lib_{base62_12chars}
-    Example: lib_2xK9mPqR7nL4
-
-    Example:
-        >>> library_id = LibraryId.generate()
-        >>> library_id.prefix
-        'lib'
-        >>> LibraryId("lib_2xK9mPqR7nL4")
-        LibraryId('lib_2xK9mPqR7nL4')
-    """
-
-    EXPECTED_PREFIX: ClassVar[str] = "lib"
-
+from src.shared_kernel.value_objects.library_id import LibraryId
 
 __all__ = ["LibraryId"]
