+++ src/shared_kernel/value_objects/__init__.py
@@ -5,6 +5,7 @@ from src.shared_kernel.value_objects.file_path import FilePath
 from src.shared_kernel.value_objects.image_url import ImageUrl
 from src.shared_kernel.value_objects.language_code import LanguageCode
 from src.shared_kernel.value_objects.language_tag import LanguageTag
+from src.shared_kernel.value_objects.library_id import LibraryId
 from src.shared_kernel.value_objects.media_type import CollectionMediaType, MediaType
 from src.shared_kernel.value_objects.profile_id import ProfileId
 from src.shared_kernel.value_objects.tracks import AudioTrack, SubtitleTrack
@@ -18,6 +19,7 @@ __all__ = [
     "ImageUrl",
     "LanguageCode",
     "LanguageTag",
+    "LibraryId",
     "MediaType",
     "ProfileId",
     "SubtitleTrack",
