+++ tests/shared_kernel/unit/value_objects/test_library_id.py
@@ -4,7 +4,7 @@ import pytest
 
 from src.building_blocks.domain.errors import DomainValidationException
 from src.building_blocks.domain.external_id import RANDOM_PART_LENGTH
-from src.modules.library.domain.value_objects.library_id import LibraryId
+from src.shared_kernel.value_objects.library_id import LibraryId
 
 
 class TestLibraryIdCreation:
